refactor: decompose analyze report into modules
